+++ pages/top.php
@@ -45,11 +45,10 @@ if($pv_meta['schema_ver'] < $db_ver){
             
             <div class="topbar-divider d-none d-sm-block"></div>
             <li class="nav-item dropdown no-arrow">
-              <a class="nav-link dropdown-toggle" href="#" id="userDropdown" role="button" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+            	<a class="nav-link dropdown-toggle" href="#" id="userDropdown" role="button" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
                 <span class="mx-2 d-none d-lg-inline text-white small"><?php echo $user['fullName'];?></span>
-               <div class="icon-container">
-                <img class="img-profile rounded-circle" src="<?=$doc['avatar']?: '/img/logo_def.png'; ?>">
-				<div class="status-circle"></div>
+               	<div class="icon-container">
+                	<img class="img-profile rounded-circle" src="<?=$doc['avatar']?: '/img/logo_def.png'; ?>">
 				</div>
               </a>
               <div class="mx-2 dropdown-menu dropdown-menu-end shadow animated--grow-in" aria-labelledby="userDropdown">
@@ -105,7 +104,7 @@ $(document).ready(function() {
 		var relUrl = 'https://raw.githubusercontent.com/globaldyne/parfumvault/master/releasenotes.md';
 	
 		$('#new-rel').load(relUrl);
-		console.log(relUrl);
+		//console.log(relUrl);
 	});
 	
 	<?php if($show_db_upgrade){?>