+++ releasenotes.md
@@ -1,3 +1,5 @@
 Whats New in v10.9
 --------------------------
+- Fix materials not updating corrrectly when added in the cart
+- Added CSV export for cart ingredients
 - This release may include more changes, for full details please refer to the CHANGELOG